🚚VideoService에 있는 mainVideo 코드를 MainService파일 생성 후 이동
